requirements.txt
# Web Framework
flask==2.3.3
flask-login==0.6.3

# Telegram Bot (исправленная версия)
aiogram==3.7.0
aiohttp==3.8.6

# Database & Storage
python-dotenv==1.0.0

# Utilities
werkzeug==2.3.7
jinja2==3.1.2
pytz==2023.3

# Security
cryptography==41.0.7

# File handling
pillow==10.0.1

# Logging (обновленная версия)
loguru==0.7.2

# Development tools
python-dateutil==2.8.2

nest_asyncio==1.5.8
requests==2.31.0
pillow
